14. Личный проект: капитан ООП
