+++ src/components/films-list.js
@@ -1,18 +1,6 @@
-import {createElement} from './../utils.js';
+import {AbstractComponent} from './abstract-component.js';
 
-export class FilmsList {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
+export class FilmsList extends AbstractComponent {
   getTemplate() {
     return `<section class="films-list">
   <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
