+++ src/components/load-more-btn.js
@@ -1,18 +1,6 @@
-import {createElement} from './../utils.js';
+import {AbstractComponent} from './abstract-component.js';
 
-export class LoadMoreBtn {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
+export class LoadMoreBtn extends AbstractComponent {
   getTemplate() {
     return `<button class="films-list__show-more">Show more</button>`;
   }
