+++ src/components/sort.js
@@ -1,18 +1,6 @@
-import {createElement} from './../utils.js';
+import {AbstractComponent} from './abstract-component.js';
 
-export class Sort {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
+export class Sort extends AbstractComponent {
   getTemplate() {
     return `<ul class="sort">
   <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
